requirements.txt
absl-py==1.0.0
attrs==21.2.0
certifi==2021.10.8
charset-normalizer==2.0.9
cycler==0.11.0
fonttools==4.28.4
idna==3.3
kiwisolver==1.3.2
matplotlib==3.5.1
mediapipe==0.8.9.1
numpy==1.21.4
opencv-contrib-python==4.5.4.60
opencv-python==4.5.4.60
packaging==21.3
Pillow==8.4.0
protobuf==3.19.1
pyparsing==3.0.6
pyserial==3.5
python-dateutil==2.8.2
python-dotenv==0.19.2
requests==2.26.0
six==1.16.0
urllib3==1.26.7
